requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Trading & Exchange APIs
pybit>=5.6.0          # Bybit API
alpaca-trade-api>=3.0.0  # Alpaca API
ccxt>=4.0.0           # Cryptocurrency Exchange Trading Library

# Data Analysis & Machine Learning
scikit-learn>=1.0.0
yfinance>=0.2.0       # Yahoo Finance data

# Async & HTTP
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
requests>=2.28.0

# Configuration & Environment
python-dotenv>=0.19.0
pydantic>=2.0.0       # Data validation
pyyaml>=6.0

# Logging & Monitoring
structlog>=22.1.0
rich>=12.0.0          # Rich console output

# Data Providers & APIs
financialmodelingprep>=0.1.0  # Financial Modeling Prep API

# Enhanced Visualization
plotly>=5.0.0         # Interactive plots
dash>=2.0.0           # Web dashboards
ipywidgets>=7.6.0     # Jupyter widgets

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.7.0

# Development
black>=22.0.0         # Code formatting
flake8>=5.0.0         # Linting
mypy>=1.0.0           # Type checking


